Now implements NAudio interfaces
The VolumeMixer implementation was very "bodge-y" and basically was a cut down version of what NAudio.Wasapi does. For this reason I have decided to opt for the well tested NAudio implementation.

This means I now have access to not just the DefaultAudioEndpoint, but all of them(EnumerateAudioEndPoints)!

This means that the blocker now operates across all audio endpoints rather than just the current one (speakers usually). So if your Spotify is set to output to a Virtual Audio Cable or something similar, then the adblocker will still see it :hatched_chick:.