Create KillerBlocker.cs
This is an advert blocker that uses an exploit in the way Spotify handles adverts when the application is restarted.

When the Spotify application exists, any playing ads are removed from the queue, and the next song that was meant to play is selected, when you relaunch the application, it primes the next song, all it takes is a press of the play button and away you go.

This IAdvertBlocker kills the Spotify processes when an advert is detected, eliminating the adverts, relaunches Spotify, it then sends a low level key event which simulates the Space bar being pressed, Spotify treats this as a press of the Play button, resuming the queue.