Update KillerBlocker.cs
Major improvement to killing logic.

When Spotify is re-launched, it is launched in minimized mode, also we no longer use SendInput, rather we send a APPCOMMAND "MEDIA_PLAY" message to the application, this has the benefit of there no longer being a race condition where the user may tab away from Spotify as the "Space" key is being virtually pressed, meaning the music does not begin. APPCOMMAND is 100% reliable.

This was originally "afk mode" due to the space bar problem, but now with the reliability of the APPCOMMAND message we treat it as the default mode.